style: :art: Move pod, node and cluster specifics to their own files
This won't actually make a difference to dependencies, it purely makes it easier to read and digest. The package remains the same (cluster), which is what's important.
